# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service C - Prasyarat dan Kurikulum API",
        "description": "API untuk validasi prasyarat dan kurikulum mahasiswa. Service ini menyediakan akses data nilai mahasiswa, IPS semester, dan detail kurikulum.",
        "contact": {
            "name": "Andi Muh. Arif Darma Saputra M",
            "email": "102022580023@student.telkomuniversity.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/nilai/{nim}": {
            "get": {
                "tags": [
                    "Nilai"
                ],
                "summary": "Lihat nilai dan IPS semester lalu berdasarkan NIM",
                "description": "Mengambil daftar nilai dan menghitung IPS (Indeks Prestasi Semester) mahasiswa. Dipanggil oleh Service B untuk cek syarat ambil matkul.",
                "operationId": "getNilaiByNim",
                "parameters": [
                    {
                        "name": "nim",
                        "in": "path",
                        "description": "NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "102022400136"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil nilai mahasiswa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data nilai mahasiswa berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "nim": {
                                                    "type": "string",
                                                    "example": "102022400136"
                                                },
                                                "ips": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 3.75
                                                },
                                                "total_sks": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "semester_terakhir": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "nilai": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "kode_matkul": {
                                                                "type": "string",
                                                                "example": "SI101"
                                                            },
                                                            "nama_matkul": {
                                                                "type": "string",
                                                                "example": "Algoritma dan Pemrograman"
                                                            },
                                                            "nilai_huruf": {
                                                                "type": "string",
                                                                "example": "A"
                                                            },
                                                            "nilai_angka": {
                                                                "type": "number",
                                                                "example": 4
                                                            },
                                                            "sks": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "semester": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/IaeT2Meta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data nilai mahasiswa tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IaeT2ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - API Key tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IaeT2ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-IAE-KEY": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "IaeT2Meta": {
                "properties": {
                    "service_name": {
                        "type": "string",
                        "example": "Prasyarat-Kurikulum-Service"
                    },
                    "api_version": {
                        "type": "string",
                        "example": "v1"
                    }
                },
                "type": "object"
            },
...
            "IaeT2ErrorResponse": {
                "required": [
                    "status",
                    "message"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Detail pesan kesalahan"
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-IAE-KEY": {
                "type": "apiKey",
                "description": "Masukkan NIM sebagai API Key (102022580023)",
                "name": "X-IAE-KEY",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer JWT dari IAE SSO (https://iae-sso.virtualfri.id/api/v1/auth/token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}